requirements.txt
openai==1.7.2
python-telegram-bot==20.7
python-dotenv==0.20.0
requests==2.31.0
beautifulsoup4==4.12.3